requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
groq==0.9.0
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
